Refactored implementations of todo list, task, database helper and data source objects. Unique ids of tables are of type string and auto generated by using UUID's. Data source methods will no longer expose the open and close methods.

+++ app/src/main/java/todo/javier/mera/todolist/database/TodoListSQLiteHelper.java
@@ -1,6 +1,7 @@
 package todo.javier.mera.todolist.database;
 
 import android.content.Context;
+import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.provider.BaseColumns;
@@ -12,39 +13,40 @@ import android.provider.BaseColumns;
 public class TodoListSQLiteHelper extends SQLiteOpenHelper {
 
     private static final String DB_NAME = "todo_lists.db";
-    private static final int DB_VERSION = 9;
+    private static final int DB_VERSION = 16;
 
     public static final String TABLE_TODO_LISTS = "TODO_LISTS";
+    public static final String COLUMN_TODO_LIST_ID = "TODOLIST_ID";
     public static final String COLUMN_TODO_LIST_NAME = "NAME";
     public static final String COLUMN_CREATION_DATE = "CREATED_DATE";
-    public static final String COLUMN_DUE_DATE = "DUE_DATE";
     public static final String COLUMN_TODO_LIST_POSITION = "POSITION";
     private String CREATE_TODO_LISTS = "CREATE TABLE "
         + TABLE_TODO_LISTS
         + "("
-        + BaseColumns._ID + " INTEGER PRIMARY KEY AUTOINCREMENT,"
+        + COLUMN_TODO_LIST_ID + " TEXT PRIMARY KEY, "
         + COLUMN_TODO_LIST_POSITION + " INTEGER, "
         + COLUMN_TODO_LIST_NAME + " TEXT, "
-        + COLUMN_CREATION_DATE + " INTEGER, "
-        + COLUMN_DUE_DATE + " INTEGER"
+        + COLUMN_CREATION_DATE + " INTEGER"
         + ")";
 
     public static final String TABLE_TODO_LIST_ITEMS = "TODO_LIST_ITEMS";
-    public static final String COLUMN_ITEMS_FOREIGN_KEY = "TODO_LIST_ID";
+    public static final String COLUMN_ITEMS_ID = "ITEM_ID";
     public static final String COLUMN_ITEMS_POSITION = "POSITION";
     public static final String COLUMN_ITEMS_DESCRIPTION = "DESCRIPTION";
     public static final String COLUMN_ITEMS_COMPLETED = "COMPLETED";
-    public static final String COLUMN_ITEMS_TIMESTAMP = "TIMESTAMP";
+    public static final String COLUMN_ITEMS_CREATED_ON = "CREATED_ON";
+    public static final String COLUMN_ITEMS_DUE_DATE = "DUE_DATE";
     private static final String CREATE_TODO_LIST_ITEMS = "CREATE TABLE "
         + TABLE_TODO_LIST_ITEMS
         + "("
-        + BaseColumns._ID + " INTEGER PRIMARY KEY AUTOINCREMENT, "
-        + COLUMN_ITEMS_FOREIGN_KEY + " INTEGER, "
+        + COLUMN_ITEMS_ID + " TEXT PRIMARY KEY, "
+        + COLUMN_TODO_LIST_ID + " TEXT, "
         + COLUMN_ITEMS_POSITION + " INTEGER, "
         + COLUMN_ITEMS_DESCRIPTION + " TEXT, "
         + COLUMN_ITEMS_COMPLETED + " INTEGER, "
-        + COLUMN_ITEMS_TIMESTAMP +  " INTEGER, "
-        + "FOREIGN KEY " + "(" + COLUMN_ITEMS_FOREIGN_KEY + ") REFERENCES " + TABLE_TODO_LISTS + "(" + BaseColumns._ID + ")"
+        + COLUMN_ITEMS_CREATED_ON +  " INTEGER, "
+        + COLUMN_ITEMS_DUE_DATE + " INTEGER, "
+        + "FOREIGN KEY " + "(" + COLUMN_TODO_LIST_ID + ") REFERENCES " + TABLE_TODO_LISTS + "(" + COLUMN_TODO_LIST_ID + ")"
         + ")";
 
     public TodoListSQLiteHelper(Context context) {
@@ -55,27 +57,25 @@ public class TodoListSQLiteHelper extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase sqLiteDatabase) {
 
-        sqLiteDatabase.execSQL(CREATE_TODO_LISTS);
-        sqLiteDatabase.execSQL(CREATE_TODO_LIST_ITEMS);
+        try {
+
+            sqLiteDatabase.execSQL(CREATE_TODO_LISTS);
+            sqLiteDatabase.execSQL(CREATE_TODO_LIST_ITEMS);
+        }
+        catch(SQLException ex) {
+
+            String message = ex.getMessage();
+        }
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion) {
 
-        switch(oldVersion) {
+        if(oldVersion < newVersion) {
 
-            case 1:
-            case 2:
-            case 3:
-            case 4:
-            case 5:
-            case 6:
-            case 7:
-            case 8:
-                sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + TABLE_TODO_LIST_ITEMS);
-                sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + TABLE_TODO_LISTS);
-                onCreate(sqLiteDatabase);
-                break;
+            sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + TABLE_TODO_LIST_ITEMS);
+            sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + TABLE_TODO_LISTS);
+            onCreate(sqLiteDatabase);
         }
     }
 }
